+++ src/main/java/calculator/processor/Calculator.java
@@ -2,7 +2,6 @@ package calculator.processor;
 
 import java.util.Optional;
 
-public interface Calculator
-{
-    Optional<Integer> calculate();
+public interface Calculator {
+    Optional<Long> calculate();
 }
